Loosen key restrictions when composing engines
Before, ObjectStorageEngine had a same-type requirement for keys. For example, an ObjectStorageEngine backed by a SQLiteStorageEngine would have to use SQLiteStorageKeys as its keys. This leaks implementation details.

Since StorageKeys now broadcast the underlying "raw value" type of the key, we can use a key-type-erasing StorageEngine to make it so that the ObjectStorageEngine can have its own key type that's distinct from the inner engine's key type, as long as both keys are backed by the same raw key type (like a `String`)